Refactored the Doubly Connected Linked List
As the core of the polygon decomposition algoiritm, the doubly-
connected linked list code really needs to be solid. Until now, it
was just a bunch of disparate objects pointing to each other that
were managed entirely by the algorithm. This change moves all of the
logic for building and maintaining that group of structures into a
a wrapper object Dcel.

The Dcel class maintains a logically-valid doubly connected linked list.
It ensures that every pointer link between edges is bi-directional,
every edge points to the same region as its neighbors and every vertex
tracks the edges that use it. In addition to maintaining the state,
the Dcel provides a number of operations that can be used to modify the
structure, expanding regions and creating new regions by splitting
existing regions or creating new edges. Which operations are supported
is directly influenced by the needs of the ccpp algorithm.

Traversal through the dcel is done through handle objects, which expose
methods to move to the linked objects. These links maintain a single
integer value referencing a specific object in the dcel, and a pointer
to the dcel. Methods on these handle objects map to direct calls on the
original dcel object that created them.

Since the management of the dcel objects and pointers was woven into
the ccpp algorithm, a bit of refactoring of the algorithm had to go
along with this. Originally, the ccpp algorithm added all edges to the
dcel beforehand, and then just split the region apart as necessary.
Now, it creates regions as it moves across the shape, extending the
existing regions by adding points as they are encountered by the sweep
line.