Separate sweep-line logic from decomposition algorithm
Having the logic to run a sweep line algorithm separate gives a little
more flexibility in where this code can be used, and more importantly,
helps keep the project from getting too highly coupled and spaghetti-y.

The sweep line algorithm is now pulled out into a separate
PolygonSweepLineAlgorithm base class that can be inherited in order
to build a specialized sweep line algorithm. The class is named
specifically 'Polygon' sweep line algorithm because it handles the sort
order of edges slightly differently than a generic sweep algorithm
would.

The Edge type is split up into two parts, a generic Edge and a
PolygonEdge that extends that type. While there is no specific use
case for the base Edge type, it could be used to write a generic
SweepLineAlgorithm base class that behaves similarly to the Polygon-
specific version.

Along with this change comes some additional splitting of the
decomposition algorithm file. It's been excessively large up to this
point, as it contained all of the code for the algorithm and the
various list types used to track edges. Now that it's pretty well known
what the necessary types are, it makes sense to do a bit of
organizational work.